[HEXDEV-692] pojo model mismatch (#1235)
* HEXDEV-692: model/pojo prediction mismatch.  Added prostate dataset and pyunit/Runit test to make sure my fix works.  Thanks to Tomas for the suggestion.

* PUBDEV-692: pojo/model prediction mismatch.  HotFix: pyunit_javapredict_dynamic_data_paramsPCA.py, wrong column number was used in generating PCA parameter K.

* HEXDEV-692: pojo/model prediction mismatch.  Reduce number of PCA models built in RotterDam.R by randomly building only two out of the four to reduce model run time.
HEXDEV-692: pojo/model prediction mismatch.  Will remove model after running R/Pyunit tests.
